day26/ai-agent-observability/backend/requirements.txt
fastapi==0.108.0
uvicorn==0.24.0
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-exporter-jaeger==1.21.0
prometheus-client==0.19.0
elasticsearch==8.11.0
redis==5.0.1
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
celery==5.3.4
google-generativeai==0.3.2
pandas==2.1.3
numpy==1.24.4
scikit-learn==1.3.2
pyyaml==6.0.1
structlog==23.2.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
websockets==12.0
aioredis==2.0.1
pydantic==2.5.0
python-multipart==0.0.6
jinja2==3.1.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
